+++ scripts/check_release_ready.sh
@@ -33,6 +33,7 @@ run_step "documentation links" ./scripts/check_docs_links.sh
 run_step "API reference coverage" ./scripts/check_api_reference.sh
 run_step "template manifest schema" ./scripts/check_template_manifest_schema.sh
 run_step "template reference table" ./scripts/check_template_reference_table.sh
+run_step "tag reference" ./scripts/check_tag_reference.sh
 run_step "MATLAB help text" ./scripts/check_matlab_help.sh
 run_step "README gallery preview" ./scripts/check_readme_gallery.sh
 run_step "toolbox independence" ./scripts/check_toolbox_independence.sh